requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
djangorestframework==3.14.0
Pillow==9.2.0
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.2
